+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `die()` will now wait 1 synchronous second before exiting
 * Fix SettingValue » GroupValue conversion in settings system
 * Make `Conduit#serveFile()` actually call the `options.onError` callback when an error occurs
+* Add `ImportStreamParser` class, deduplicating logic in both alchemy.js and model.js
 
 ## 1.4.0 (2026-01-21)